+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. This change
 ## [2-beta28] - unreleased
 - added a `snippets` API namespace for easier editor inntegration - by @timothyprately
 - added a `kindly-compute` endpoint - for requesting server computations - with @SpacePirate27
+- added `kind/scittle`
+- added automatic recognition of `kind/scittle` and `kind/reagent` when nested inside `kind/hiccup`
+- added basic support for a Scittle header script with initial setup
 
 ## [2-beta27] - 2024-12-24
 - added a `:browse` option (default `true`) to determine whether to open a browser tab when the server is started